Fix priorities for tasks
When setting and changing priorities no attention was paid to the
`story_id', which means that no two tasks could have the same priority,
even when belonging to other stories. This creates a weird situation
where the tasks of a certain story would become unmanageable because
there are gaps between the priorities. This could make it seem like
priority changing was impossible.